Questions .yes_or_no? .multichoice
Questions.yes_or_no?  (string) -> bool
takes a question and asks the users yes or no.
It reruns true or flase

Questions.multichoice (string, array) -> item from array
taks a questions and asks the user to pick one of the items in the array
by number. It then returns the item in the array