Questions .yes_or_no? .multichoice (#33)
* Questions .yes_or_no? .multichoice

Questions.yes_or_no?  (string) -> bool
takes a question and asks the users yes or no.
It reruns true or flase

Questions.multichoice (string, array) -> item from array
taks a questions and asks the user to pick one of the items in the array
by number. It then returns the item in the array

* Replaced validate_user_input with new methods/ updated tests

Co-authored-by: Tyler Decker <c_tydecker@groupon.com>
Co-authored-by: Cristian-Baeza <cristianbaeza092@gmail.com>